Move music player code its own folder
