+++ cogs/nodemanager.py
@@ -5,7 +5,6 @@ from discord.ext import commands
 from lavalink import Node
 
 import codecs
-
 import yaml
 
 # Bot Utilities
