+++ cogs/utils/mixplayer/__init__.py
@@ -1,0 +1,3 @@
+from .player import MixPlayer
+
+__all__ = ['MixPlayer']
